+++ gonggibap/src/types/restaurant.ts
@@ -1,6 +1,6 @@
 export interface GetRestaurantsResponse {
   totalPages: number;
-  restaurantResponses: Restaurant[];
+  content: Restaurant[];
 }
 
 export interface Restaurant {
@@ -8,6 +8,7 @@ export interface Restaurant {
   restaurantName: string;
   restaurantLink: string;
   restaurantCategory: string;
+  restaurantDetailCategory: string;
   restaurantAddressName: string;
   restaurantRoadAddressName: string;
   restaurantLatitude: number;
